+++ examples/createSellerFunnelBase.php
@@ -16,6 +16,9 @@
 
     $createSellerFunnelBase->platform_id = "123";
     $createSellerFunnelBase->shop_url = "http://localhost";
+    $createSellerFunnelBase->platform_version = "1.0.0";
+    $createSellerFunnelBase->plugin_version = "1.0.0";
+    $createSellerFunnelBase->site_id = "MLB";
 
     debug(json_encode($createSellerFunnelBase->save()));
  ?>