+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## V2.8.0
+### Added
+- Is possible using empty constructor on Sdk
+- Add new fields on seller base funnel: site_id, platform_version, plugin_version and is_disabled
+
 ## V2.7.0
 ### Added
 - Added `id`, `status`, `payment_type_id` parameters on Payment response